products endpoints (get, post, put, delete) ajouter

+++ web-version/backend/controllers/stocks-controller.js
@@ -28,10 +28,10 @@ const addStock = async (req, res, next) => {
 
     // check authorization req.userData.id (from the token)
 
-    const { title, typeStock, description } = req.body;
+    const { title, typeStock, description, restaurantId } = req.body;
 
-    Stock.create({ title, type_stock: typeStock, description }).then(response => {
-        res.status(200).json({ message: "Stock created...", title, typeStock, description });
+    Stock.create({ title, type_stock: typeStock, description, RestaurantId: +restaurantId }).then(response => {
+        res.status(200).json({ message: "Stock created...", title, typeStock, description, restaurantId });
     }).catch(error => {
         console.log(error.message);
         res.status(404).json({ error: "Un error occured" });
@@ -43,9 +43,9 @@ const updateStock = async (req, res, next) => {
 
     // check authorization req.userData.id (from the token) 
 
-    const { id, title, typeStock, description } = req.body;
+    const { id, title, typeStock, description, restaurantId } = req.body;
 
-    Stock.update({ title, type_stock: typeStock, description }, { where: { id }})
+    Stock.update({ title, type_stock: typeStock, description, RestaurantId: +restaurantId }, { where: { id }})
             .then( () => {
                 res.status(201).json({
                     message: `Stock with id ${id} updated...`
@@ -64,18 +64,24 @@ const deleteStock = async (req, res, next) => {
 
     // check authorization req.userData.id (from the token)
 
-    const id = req.params['id'];
+    const id = +req.params['id'];
+
+    if(!Number.isInteger(id)) {
+        return res.status(404).json({ error: "please provide a valid integer as a stockId" });
+    }
+
     Stock.destroy({ where: { id: id}})
         .then(response => {
+            console.log(response);
             res.status(201).json({
                 message: `Stock with id ${id} deleted...`
-            })
+            });
         }).catch(error => {
             console.log(error.message);
-            res.status(404).json({
-                message: "Un error occured"
+            res.status(500).json({
+                message: "Un error occured",
             })
-        })
+        });
 }
